backstage: generate components for o-buttons
