--- components/o3-buttons/src/css/core/_variables.css
@@ -1,26 +1,0 @@
-/**
- * Do not edit directly
- * Generated on Mon, 21 Aug 2023 14:29:13 GMT
- */
-
-
-.o-brand-core .o-buttons-experimental {
-  --_o-primary-active-background-color: red;
-  --_o-primary-active-color: var(--o-color-usecase-button-primary-active);
-  --_o-primary-active-border: 1px solid var(--o-color-usecase-background-button-primary-active);
-  --_o-primary-hover-background-color: var(--o-color-usecase-background-button-primary-hover);
-  --_o-primary-hover-color: var(--o-color-usecase-button-primary-hover);
-  --_o-primary-hover-border: 1px solid var(--o-color-usecase-background-button-primary-hover);
-  --_o-secondary-active-background-color: var(--o-color-usecase-background-button-secondary-active);
-  --_o-secondary-active-color: var(--o-color-usecase-button-secondary-active);
-  --_o-secondary-active-border: 1px solid var(--o-color-usecase-background-button-primary-active);
-  --_o-secondary-hover-background-color: var(--o-color-usecase-background-button-secondary-hover);
-  --_o-secondary-hover-color: var(--o-color-usecase-button-secondary-hover);
-  --_o-secondary-hover-border: 1px solid var(--o-color-usecase-background-button-primary-active);
-  --_o-ghost-active-background-color: var(--o-color-usecase-background-button-secondary-active);
-  --_o-ghost-active-color: var(--o-color-usecase-button-secondary-active);
-  --_o-ghost-active-border: none;
-  --_o-ghost-hover-background-color: var(--o-color-usecase-background-button-secondary-hover);
-  --_o-ghost-hover-color: var(--o-color-usecase-button-secondary-hover);
-  --_o-ghost-hover-border: none;
-}
