+++ components/o3-buttons/src/css/internal/o3-buttons/_variables.css
@@ -4,7 +4,7 @@
  */
 
 
-.o-brand-core .o3-buttons {
+.o-brand-internal .o3-buttons {
   --_o3-buttons-primary-standard-color: #ffffff;
   --_o3-buttons-primary-standard-background: var(--o3-color-palette-teal);
   --_o3-buttons-primary-standard-border: #ffffff00;
