+++ components/o3-buttons/stories/buttons-professional.scss
@@ -3,3 +3,4 @@
 @include oFonts();
 @include oNormalise();
 
+@import '../main.css';
