+++ components/o3-buttons/stories/buttons.scss
@@ -7,11 +7,8 @@
 
 @import '../main.css';
 
-.o3-buttons, .o3-buttons--primary {
-  // background-color: var(--o-background-color-root);
-  // color: var(--o-color-button-primary-active);
-	color: var(--o3-red)
-}
+
+
 
 .sub-container {
   background-color: var(--o-background-color-root);
